# web/swagger.json
{
  "openapi" : "3.0.0",
  "servers" : [
    {"url" : "http://localhost:8001/api/v1"}
  ],
  "info" : {
    "version" : "1.0.0-oas3",
    "title" : "Swagger Koa Skeleton API"
  },
...
  "paths" : {
...
    "/forgot-password" : {
      "post" : {
        "tags" : [ "Forgot password" ],
        "summary" : "",
        "operationId" : "forgotPassword",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ForgotPasswordRequest"
              }
            }
          },
          "description" : "Forgot password data",
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Entity is created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request"
          },
          "422" : {
            "description" : "Faild validation"
          },
          "500" : {
            "description" : "Server error"
          }
        }
      },
      "patch" : {
        "tags" : [ "Forgot password" ],
        "summary" : "",
        "operationId" : "patchForgotPassword",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PatchForgotPasswordRequest"
              }
            }
          },
          "description" : "Forgot password data",
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Password is updated",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request"
          },
          "422" : {
            "description" : "Faild validation"
          },
          "500" : {
            "description" : "Server error"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ForgotPasswordRequest" : {
        "type" : "object",
        "required" : [ "email" ],
        "properties" : {
          "email" : {
            "type" : "string",
            "example" : "test@gmail.com"
          }
        }
      },
      "PatchForgotPasswordRequest" : {
        "type" : "object",
        "required" : [ "token" ],
        "properties" : {
          "token" : {
            "type" : "string",
            "example" : "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6NSwiZW1haWwiOiJ0ZWFjaGVyQGdtYWlsLmNvbSIsInJvbGUiOiJ0ZWFjaGVyIiwiZmlyc3ROYW1lIjoiVGVhY2hlciIsImxhc3ROYW1lIjoiTGFzdCBuYW1lIiwiZGVzY3JpcHRpb24iOm51bGwsImlhdCI6MTUxODk1NTI3M30.R7-jnh5lkQpiCjBmnke1NTfzs_onyoVTRu1zjr2E0XQ"
          },
          "password" : {
            "type" : "string",
            "example" : "1234abCD"
          }
        }
      },
...
      "SuccessResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean",
            "example" : true
          }
        }
      }
    },
    "securitySchemes" : {
      "mspSecurity" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}